Move test to the right directory and add license text
