+++ src/tests/gov/nasa/jpf/test/java/util/StreamTest.java
@@ -1,0 +1,41 @@
+/*
+ * Copyright (C) 2014, United States Government, as represented by the
+ * Administrator of the National Aeronautics and Space Administration.
+ * All rights reserved.
+ *
+ * The Java Pathfinder core (jpf-core) platform is licensed under the
+ * Apache License, Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License. You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0.
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package gov.nasa.jpf.test.java.util;
+
+import gov.nasa.jpf.util.test.TestJPF;
+
+import java.util.Arrays;
+import java.util.List;
+import java.util.stream.Collectors;
+import org.junit.Test;
+import static java.util.Arrays.asList;
+
+public class StreamTest extends TestJPF {
+    public static List<String> allToUpperCase(List<String> words) {
+        return words.stream()
+                    .map(string -> string.toUpperCase())
+                    .collect(Collectors.toList());
+    }
+
+    @Test
+public void multipleWordsToUppercase() {
+    List<String> input = Arrays.asList("a", "b", "hello");
+    List<String> result = allToUpperCase(input);
+    assertEquals(asList("A", "B", "HELLO"), result);
+    }
+}
